refactor, feat: reorganize gui imports and add speech recognition functionality
