+++ setting/setting.json
@@ -1,8 +1,9 @@
 {
-    "discord_RPC": true,
+    "discord_RPC": false,
     "Gesture-Control": false,
     "Gesture-Control-Screen": false,
-    "volume": 40,
+    "STT-Contorl": true,
+    "volume": 10,
     "file-save-dir": "./asset/storage",
     "Subtitle-Lang": "ko"
 }
